cab-share/offchain-crypto/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pycryptodome>=3.18.0
cryptography>=41.0.0
python-dotenv>=1.0.0
gunicorn==21.2.0
